# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Auth
    description: Browser authentication operations.
  - name: Rooms
    description: Protected room foundation for future multiplayer support.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Authentication state resolved successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login.
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        "302":
          description: Redirect to the validated return path after login.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: "#/components/parameters/SessionCookie"
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to the OIDC provider logout endpoint.
  /rooms:
    post:
      tags: [Rooms]
      operationId: createRoom
      summary: Create a protected room for the authenticated user
      parameters:
        - $ref: "#/components/parameters/SessionCookie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room and host membership created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetails"
        "400":
          description: Invalid room data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rooms/join:
    post:
      tags: [Rooms]
      operationId: joinRoom
      summary: Join a waiting room by room code
      parameters:
        - $ref: "#/components/parameters/SessionCookie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinRoomInput"
      responses:
        "200":
          description: Membership created or already present.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetails"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room not found or user is not a member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Room state does not allow this action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rooms/{roomId}/leave:
    post:
      tags: [Rooms]
      operationId: leaveRoom
      summary: Leave a room; host leaving transfers host
      parameters:
        - $ref: "#/components/parameters/SessionCookie"
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Room after the player left.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetails"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room not found or user is not a member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rooms/{roomId}/selection:
    post:
      tags: [Rooms]
      operationId: setRoomSelection
      summary: Select the caller's character and/or career before the match starts
      parameters:
        - $ref: "#/components/parameters/SessionCookie"
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectionInput"
      responses:
        "200":
          description: Updated room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetails"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room not found or user is not a member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Room state does not allow this action, or the pick is taken.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rooms/{roomId}/match:
    get:
      tags: [Match]
      operationId: getRoomMatch
      summary: Get the canonical match state (also the caller's heartbeat and reconnect recovery)
      parameters:
        - $ref: "#/components/parameters/SessionCookie"
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Current authoritative match snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchSnapshot"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room not found, not a member, or match not started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rooms/{roomId}/match/actions:
    post:
      tags: [Match]
      operationId: submitMatchAction
      summary: Request a game action; the server decides whether it is legal
      parameters:
        - $ref: "#/components/parameters/SessionCookie"
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchActionInput"
      responses:
        "200":
          description: Action accepted; the new authoritative snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchSnapshot"
        "400":
          description: Malformed action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: It is not the caller's turn.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room not found, not a member, or match not started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Illegal action, stale version, or match over.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rooms/{roomId}/ready:
    post:
      tags: [Rooms]
      operationId: setRoomReady
      summary: Set the caller's ready state
      parameters:
        - $ref: "#/components/parameters/SessionCookie"
        - $ref: "#/components/parameters/RoomId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReadyInput"
      responses:
        "200":
          description: Updated room.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetails"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room not found or user is not a member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Room state does not allow this action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rooms/{roomId}/start:
    post:
      tags: [Rooms]
      operationId: startRoom
      summary: Host-only start of the match
      parameters:
        - $ref: "#/components/parameters/SessionCookie"
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Room moved to in_progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetails"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Action not permitted for this user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room not found or user is not a member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Room state does not allow this action.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rooms/{roomId}:
    get:
      tags: [Rooms]
      operationId: getRoom
      summary: Retrieve a room only as an active or reconnecting member
      parameters:
        - $ref: "#/components/parameters/RoomId"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Room and player records.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetails"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room not found or user is not a member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /rooms/{roomId}/settings:
    patch:
      tags: [Rooms]
      operationId: updateRoomSettings
      summary: Update room settings as the host while the room is waiting
      parameters:
        - $ref: "#/components/parameters/RoomId"
        - $ref: "#/components/parameters/SessionCookie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomSettingsInput"
      responses:
        "200":
          description: Updated room and player records.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomDetails"
        "400":
          description: Invalid room settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Only the room host can change settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Room not found or user is not a member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Room settings cannot change after the room starts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  parameters:
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: HttpOnly browser session cookie.
      schema:
        type: string
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, "null"]
          format: email
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        profileImageUrl:
          type: [string, "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
    RoomStatus:
      type: string
      enum: [waiting, in_progress, completed, cancelled]
    RoomPlayerStatus:
      type: string
      enum: [joined, disconnected, left]
    Room:
      type: object
      required: [id, code, hostUserId, boardId, mode, status, minPlayers, maxPlayers, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        minPlayers:
          type: integer
          minimum: 2
          maximum: 4
        maxPlayers:
          type: integer
        hostUserId:
          type: string
        boardId:
          type: string
        mode:
          type: string
        status:
          $ref: "#/components/schemas/RoomStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RoomPlayer:
      type: object
      required: [userId, displayName, slot, status, selectedCharacterId, selectedCareerId, ready, joinedAt, updatedAt]
      properties:
        userId:
          type: string
        displayName:
          type: string
        slot:
          type: integer
          minimum: 0
          maximum: 3
        status:
          $ref: "#/components/schemas/RoomPlayerStatus"
        selectedCharacterId:
          type: [string, "null"]
        selectedCareerId:
          type: [string, "null"]
        ready:
          type: boolean
        joinedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RoomDetails:
      type: object
      required: [room, players]
      properties:
        room:
          $ref: "#/components/schemas/Room"
        players:
          type: array
          items:
            $ref: "#/components/schemas/RoomPlayer"
    RoomInput:
      type: object
      required: [boardId, mode]
      properties:
        boardId:
          type: string
          minLength: 1
          maxLength: 80
        mode:
          type: string
          minLength: 1
          maxLength: 40
    JoinRoomInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 4
          maxLength: 12
    SelectionInput:
      type: object
      minProperties: 1
      properties:
        characterId:
          type: string
          minLength: 1
          maxLength: 80
        careerId:
          type: string
          minLength: 1
          maxLength: 80
    MatchActionInput:
      type: object
      required: [action]
      properties:
        expectedVersion:
          type: integer
          minimum: 0
          description: Optional; the action is rejected with 409 if the state has moved on.
        action:
          type: object
          required: [type]
          additionalProperties: true
          description: Client intent (ROLL_DICE, MOVE, DRAW_CARD, USE_ABILITY, BUY_ASSET, USE_UPGRADE_TOKEN, GAMBLE, CHANGE_CAREER, END_TURN, ...).
          properties:
            type:
              type: string
              minLength: 1
              maxLength: 40
    MatchSeat:
      type: object
      required: [userId, playerId, playerIndex, displayName, connected]
      properties:
        userId:
          type: string
        playerId:
          type: string
        playerIndex:
          type: integer
        displayName:
          type: string
        connected:
          type: boolean
    MatchSnapshot:
      type: object
      required: [roomId, version, you, currentPlayerIndex, currentUserId, seats, match]
      properties:
        roomId:
          type: string
          format: uuid
        version:
          type: integer
        you:
          type: object
          required: [userId, playerId, playerIndex]
          properties:
            userId:
              type: string
            playerId:
              type: string
            playerIndex:
              type: integer
        currentPlayerIndex:
          type: integer
        currentUserId:
          type: [string, "null"]
        seats:
          type: array
          items:
            $ref: "#/components/schemas/MatchSeat"
        match:
          type: object
          additionalProperties: true
          description: The canonical Match (same shape as the single-player reducer state, including eventLog).
    ReadyInput:
      type: object
      required: [ready]
      properties:
        ready:
          type: boolean
    RoomSettingsInput:
      type: object
      minProperties: 1
      properties:
        minPlayers:
          type: integer
          minimum: 2
          maximum: 4
        boardId:
          type: string
          minLength: 1
          maxLength: 80
        mode:
          type: string
          minLength: 1
          maxLength: 40
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string