# lib/api-zod/src/generated/types/matchSeat.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MatchSeat {
  userId: string;
  playerId: string;
  playerIndex: number;
  displayName: string;
  connected: boolean;
}