# lib/api-zod/src/generated/types/matchSnapshot.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MatchSeat } from './matchSeat';
import type { MatchSnapshotMatch } from './matchSnapshotMatch';
import type { MatchSnapshotYou } from './matchSnapshotYou';

export interface MatchSnapshot {
  roomId: string;
  version: number;
  you: MatchSnapshotYou;
  currentPlayerIndex: number;
  /** @nullable */
  currentUserId: string | null;
  seats: MatchSeat[];
  /** The canonical Match (same shape as the single-player reducer state, including eventLog). */
  match: MatchSnapshotMatch;
}