# lib/api-zod/src/generated/types/roomPlayer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoomPlayerStatus } from './roomPlayerStatus';

export interface RoomPlayer {
  userId: string;
  displayName: string;
  /**
     * @minimum 0
     * @maximum 3
     */
  slot: number;
  status: RoomPlayerStatus;
  /** @nullable */
  selectedCharacterId: string | null;
  /** @nullable */
  selectedCareerId: string | null;
  ready: boolean;
  joinedAt: Date;
  updatedAt: Date;
}